fix: fix app rendering on locale dir
